Move status checks to a dedicated module. Remove multithread flavour
