+++ crates/shielder-relayer/tests/component_tests.rs
@@ -1,46 +1,17 @@
 use parameterized::parameterized;
-use reqwest::{Response, StatusCode};
-use rust_decimal::Decimal;
+use reqwest::{ StatusCode};
 use shielder_account::Token;
-use shielder_relayer::{RelayResponse, SimpleServiceResponse};
+use shielder_relayer::RelayResponse;
 
 use crate::utils::{
     config::{ShielderContract, TestConfig},
-    container_logs, response_message, TestContext, ERC20_ADDRESS,
+    container_logs, response_message, simple_payload, TestContext, ERC20_ADDRESS,
 };
 
 mod utils;
 
-async fn simple_payload(response: Response) -> String {
-    response_message::<SimpleServiceResponse>(response)
-        .await
-        .message
-}
-
-#[tokio::test(flavor = "multi_thread")]
-async fn in_correct_setting_service_is_healthy_and_signers_have_funds() {
-    let test_context = TestContext::default().await;
-
-    let health_response = test_context.reach_health().await;
-    ctx_assert!(health_response.status().is_success(), test_context);
-    ctx_assert_eq!(
-        simple_payload(health_response).await,
-        "Healthy",
-        test_context
-    );
-
-    let metrics = test_context.get_metrics().await;
-    ctx_assert!(
-        metrics.contains(&format!(
-            "signer_balances{{address=\"{}\"}} 20",
-            test_context.signer.address()
-        )),
-        test_context
-    );
-}
-
 #[parameterized(token = { Token::Native, Token::ERC20(ERC20_ADDRESS) })]
-#[parameterized_macro(tokio::test(flavor = "multi_thread"))]
+#[parameterized_macro(tokio::test)]
 async fn relay_query_without_quote_before_fails(token: Token) {
     let context = TestContext::default().await;
     let response = context.relay(Default::default(), token).await;
@@ -54,16 +25,7 @@ async fn relay_query_without_quote_before_fails(token: Token) {
 }
 
 #[parameterized(token = { Token::Native, Token::ERC20(ERC20_ADDRESS) })]
-#[parameterized_macro(tokio::test(flavor = "multi_thread"))]
-async fn server_returns_quotation(token: Token) {
-    let context = TestContext::default().await;
-    let quote = context.quote(token).await;
-    assert_eq!(quote.native_token_unit_price, Decimal::new(1, 18));
-    assert_eq!(quote.fee_token_unit_price, Decimal::new(1, 18));
-}
-
-#[parameterized(token = { Token::Native, Token::ERC20(ERC20_ADDRESS) })]
-#[parameterized_macro(tokio::test(flavor = "multi_thread"))]
+#[parameterized_macro(tokio::test)]
 async fn when_contract_returns_ok_server_sings_success(token: Token) {
     let test_context = TestContext::default().await;
     let response = test_context.relay_with_quote(token).await;
@@ -73,7 +35,7 @@ async fn when_contract_returns_ok_server_sings_success(token: Token) {
 }
 
 #[parameterized(token = { Token::Native, Token::ERC20(ERC20_ADDRESS) })]
-#[parameterized_macro(tokio::test(flavor = "multi_thread"))]
+#[parameterized_macro(tokio::test)]
 async fn when_contract_reverts_server_screams_failure(token: Token) {
     let config = TestConfig {
         shielder_contract: ShielderContract::Reverting,
@@ -91,7 +53,7 @@ async fn when_contract_reverts_server_screams_failure(token: Token) {
     );
 }
 
-#[tokio::test(flavor = "multi_thread")]
+#[tokio::test]
 async fn metrics_register_withdrawals() {
     let context = TestContext::default().await;
